fw-update: Improve config matching logic for EID comparisons
The previous implementation had issues matching configurations due to
EID being of type uint8_t. The comparison was failing since the value in
the MCTP interface map was implicitly being converted to int32_t instead
of uint8_t

Tests:
- Firmware Inventory populated on Bianca BMC

Fixes Jira https://develop-next-sanity

+++ requester/mctp_endpoint_discovery.cpp
@@ -362,7 +362,7 @@ void MctpDiscovery::refreshEndpoints(sdbusplus::message::message& msg)
                         "xyz.openbmc_project.Common.UUID");
 
                 const auto eid =
-                    pldm::utils::DBusHandler().getDbusProperty<uint32_t>(
+                    pldm::utils::DBusHandler().getDbusProperty<uint8_t>(
                         objPath.c_str(), "EID",
                         "xyz.openbmc_project.MCTP.Endpoint");
